Renomeia página de gestão de usuários
